+++ tomcat/src/test/java/org/apache/coyote/http11/request/requestline/HttpRequestLineTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.request.requestline;
 
 import static support.FakeRequests.invalidGetRequest;
 import static support.FakeRequests.validGetRequest;
